Rename initialRatings temporarily
