+++ src/components/Review/initRatings.js
